extracting more things from the large rendering files into separate translation units
